services/pdf-renderer/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-cloud-storage==3.2.0
PyMuPDF==1.26.4
